♻️ (refactor): Simplify import paths and enhance logging messages; update feature calculator attributes for consistency
