+++ pyproject.toml
@@ -65,6 +65,7 @@ testpaths = ["tests"]
 python_files = ["test_*.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
+pythonpath = ["src"]
 addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
 
 [tool.mypy]
