+++ src/features/calculations.py
@@ -26,10 +26,6 @@ class FeatureCalculator:
     DEFAULT_SECONDS_PER_REP = 4  # Temps moyen par répétition en secondes
     DEFAULT_SET_REST_TIME = 60   # Temps de récupération moyen entre séries
     
-    # Attributs de classe pour compatibilité avec les tests
-    SECONDS_PER_REP = 4.0  # Temps moyen par répétition en secondes
-    SET_REST_TIME = 60.0   # Temps de récupération moyen entre séries
-    
     def __init__(self, seconds_per_rep: Optional[float] = None, set_rest_time: Optional[float] = None):
         """
         Initialise les calculateurs spécialisés.
